two-way data binding for selectedDate
This patch enables two-way data binding for the 'selectedDate' attribute.
It accepts any valid date string as input for the selected date,
but returns dates in 'yyyy-MM-dd' format.
The user may still use the 'date' attribute if s/he needs to obtain the
selected date as a Javascript Date object (output only).