User Clicks and Min/Max Support
f a user clicks a date number, a flag is now set. This helps to
distuingish states of user interaction vs. programmatically setting the
date and reacting on it in the d-calendar-date-selected event handler.

Further by setting the minDate and/or the maxDate, certain minimum and
maximum can be set so the user cannot select dates outside of this
range.